Avoid deadlock in `__afl_evocatio_GetCapability`
According to some user feedbacks, we have learned that if `__asan_locate_address` is called to check an `invalidAddr` which points to sth on the stack or globally, DEADLOCK will arise.
We have carefully investigated this and you can view https://github.com/llvm/llvm-project/issues/61860 for more details.
You can use https://github.com/SonicStark/microBug to reproduce or verify this DEADLOCK.

However days have passed since llvm/llvm-project#61860 was opened, and no response has yet been seen.
To completely avoid the risk of deadlock, it is recommended that
1.  Update your Evocatio with this commit.
    According to https://github.com/llvm/llvm-project/blob/b5c862e15caf4d8aa34bbc6ee25af8da9a9405a4/compiler-rt/lib/asan/asan_debugging.cpp#L128 ,
    our `objectType` should be "low shadow", "shadow gap", "high shadow", "heap", "stack", "global" or "heap-invalid".
    So `bugT` semantically contains the info in some degree, and `objectType` can be deprecated then.

2.  If you indeed depend on `objectType`, build a special AddressSanitizer runtime from LLVM source code with those things modified as llvm/llvm-project#61860 mentioned, and replace the runtime in your current compiler tool chain. We do not provide any futher guidance on this.