Use fully-qualified paths
When afl-fuzz informs target of the path to .cur_input besides the file descriptor, it writes the path string in `detect_file_args`. Here fully-qualified path is used. So it should be better to apply the same stuff, i.e. be sure that we're always using fully-qualified path, for pCapResFilePath.