Merge by reason of A bug in the bytes-identifier (#4)
User [HotSpurzzZ](https://github.com/HotSpurzzZ) found that when some bytes are deduced as 'is_non_crashing_byte', if they are deduced as C-D byte, the final func_lab value may be greater than 7. This is a bug.

Co-authored-by: root <root@xingtu02.yjy.zzyc.qianxin-inc.cn>

It seems that the commit msgs use host name and user account as `user.email` instead of a valid one. We choose to squash and merge so that tracking the modification can be easier.